WIP: restructure so we are doung both ug and pg deliveries

+++ application/tasks/sitsdbimport.php
@@ -66,62 +66,65 @@ class SITSDBImport_Task
 
 		Log::$logfile = path('storage') . '/logs/sits_db_import.log';
 		Log::purge();
-
-		$data = $this->load_sitsimport_data('pg', $this->processYears['pg']);
+		
+		foreach ($this->processYears as $level => $year) {
+			$data[$level] = $this->load_sitsimport_data($level, $year);
+			// code...
+		}
 		// if we have no data, exit
 		if (!$data) {
 			exit;
 		}
 
-		$courseGroupings = [];
-		foreach ($data as $delivery) {
-				//only add delivery if it is new or has an MCR that is lower than the previous one
-				if (
-					isset($courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1]) &&
-					intval($delivery->sits_apply_link_code2) > intval($courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1]->sits_apply_link_code2)
-				) {
+		foreach ($data as $level => $deliveries) {
+			// code...
+			$courseGroupings = [];
+			foreach ($data as $delivery) {
+					//only add delivery if it is new or has an MCR that is lower than the previous one
+					if (
+						isset($courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1]) &&
+						intval($delivery->sits_apply_link_code2) > intval($courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1]->sits_apply_link_code2)
+					) {
+						continue;
+					}
+
+					$courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1] = $delivery;
+			}
+		
+			$this->purgeOldData($level, $this->processYears[$level]);
+			
+			foreach ($courseGroupings as $programme_id => $deliveries) {
+
+				$firstDelivery = null;
+				foreach ($deliveries as $mcr => $delivery) {
+				 	$firstDelivery = $delivery;
+				 	break;
+				 }
+				
+			  //get rid of 'UG'/'PG' that SITS concat to our progsplant ID
+				$programme = $this->getProgramme($firstDelivery->pp_page_id, $level);
+
+				if (empty($programme) || !is_object($programme)) {
 					continue;
 				}
 
-				$courseGroupings[$delivery->pp_page_id][$delivery->sits_apply_link_code1] = $delivery;
-		}
-
-		// If XML file is good, purge old caches before starting import
-		$this->purgeOldPGData($this->processYears['pg']);
-		$this->purgeOldUGData($this->processYears['ug']);
-		
-		foreach ($courseGroupings as $programme_id => $deliveries) {
-
-			$firstDelivery = null;
-			foreach ($deliveries as $mcr => $delivery) {
-			 	$firstDelivery = $delivery;
-			 	break;
-			 }
-			
-		  //get rid of 'UG'/'PG' that SITS concat to our progsplant ID
-			$level = strtolower(substr($firstDelivery->sits_study_level, 0, 2)); // or pp_award_level?
-			$programme = $this->getProgramme($firstDelivery->pp_page_id, $level);
-
-			if (empty($programme) || !is_object($programme)) {
+				echo $programme->instance_id . " =>". PHP_EOL . implode(PHP_EOL, array_map(function($delivery){
+					return $delivery->sits_academic_year 
+							. " " . $delivery->sits_study_level 
+							. " " . $delivery->sds_pos_code 
+							. " " . $delivery->sits_apply_link_code1 
+							. " " . $delivery->sits_apply_link_code2 
+							. " " . $delivery->pp_award_id 
+							. " " . $delivery->pp_discover_uni_id;
+				}, $deliveries)) . PHP_EOL;
 				continue;
-			}
 
-			echo $programme->instance_id . " =>". PHP_EOL . implode(PHP_EOL, array_map(function($delivery){
-				return $delivery->sits_academic_year 
-						. " " . $delivery->sits_study_level 
-						. " " . $delivery->sds_pos_code 
-						. " " . $delivery->sits_apply_link_code1 
-						. " " . $delivery->sits_apply_link_code2 
-						. " " . $delivery->pp_award_id 
-						. " " . $delivery->pp_discover_uni_id;
-			}, $deliveries)) . PHP_EOL;
-			continue;
-
-			$this->updateKISCourseID($programme, $firstDelivery->pp_discover_uni_id);
-			
-			URLParams::$type = $level;
-			foreach ($deliveries as $mcr => $deliveryData) {
-				$delivery = $this->createDelivery($programme, $deliveryData);
+				$this->updateKISCourseID($programme, $firstDelivery->pp_discover_uni_id);
+				
+				URLParams::$type = $level;
+				foreach ($deliveries as $mcr => $deliveryData) {
+					$delivery = $this->createDelivery($programme, $deliveryData);
+				}
 			}
 		}
 	
@@ -170,10 +173,22 @@ class SITSDBImport_Task
   /**
    * Remove all PG deliveries
    */
-	public function purgeOldPGData($year)
+	public function purgeOldData($level, $year)
 	{
-		$to_del = DB::table('programmes_pg')->where('year', '=', $year)->lists('id');
-		DB::table('pg_programme_deliveries')->where_in('programme_id', $to_del)->delete();
+		if ($level === 'pg') {
+			$to_del = DB::table('programmes_pg')->where('year', '=', $year)->lists('id');
+			if (!empty($to_del)) {
+				DB::table('pg_programme_deliveries')->where_in('programme_id', $to_del)->delete();
+			}
+		}
+		
+		if ($level === 'ug') {
+			$to_del = DB::table('programmes_ug')->where('year', '=', $year)->lists('id');
+			if (!empty($to_del)) {
+				DB::table('ug_programme_deliveries')->where_in('programme_id', $to_del)->delete();
+			}
+		}
+
 	}
 
   /**
@@ -181,8 +196,7 @@ class SITSDBImport_Task
    */
 	public function purgeOldUGData($year)
 	{
-		$to_del = DB::table('programmes_ug')->where('year', '=', $year)->lists('id');
-		DB::table('ug_programme_deliveries')->where_in('programme_id', $to_del)->delete();
+		
 	}
 
   /**
